Move managers to managers folder
