+++ src/cryptoadvance/specter/managers/device_manager.py
@@ -1,12 +1,12 @@
 import os
 import json
 import logging
-from .helpers import alias, load_jsons
-from .rpc import get_default_datadir
+from ..helpers import alias, load_jsons
+from ..rpc import get_default_datadir
 
-from .devices import __all__ as device_classes
-from .devices.generic import GenericDevice  # default device type
-from .persistence import write_device, delete_file, delete_folder
+from ..devices import __all__ as device_classes
+from ..devices.generic import GenericDevice  # default device type
+from ..persistence import write_device, delete_file, delete_folder
 
 logger = logging.getLogger(__name__)
