+++ tests/conftest.py
@@ -13,7 +13,7 @@ from cryptoadvance.specter.bitcoind import (
     BitcoindPlainController,
 )
 from cryptoadvance.specter.bitcoind_docker import BitcoindDockerController
-from cryptoadvance.specter.device_manager import DeviceManager
+from cryptoadvance.specter.managers.device_manager import DeviceManager
 from cryptoadvance.specter.specter import Specter
 from cryptoadvance.specter.server import create_app, init_app
