+++ tests/test_specter.py
@@ -5,7 +5,7 @@ from cryptoadvance.specter.key import Key
 from cryptoadvance.specter.rpc import BitcoinRPC
 from cryptoadvance.specter.specter import get_rpc, Specter
 from cryptoadvance.specter.specter_error import SpecterError
-from cryptoadvance.specter.wallet_manager import WalletManager
+from cryptoadvance.specter.managers.wallet_manager import WalletManager
 
 
 def test_alias():
