+++ tests/test_wallet_manager.py
@@ -4,7 +4,7 @@ from cryptoadvance.specter.rpc import RpcError
 from cryptoadvance.specter.specter_error import SpecterError
 from cryptoadvance.specter.wallet import Wallet
 from cryptoadvance.specter.key import Key
-from cryptoadvance.specter.wallet_manager import WalletManager
+from cryptoadvance.specter.managers.wallet_manager import WalletManager
 
 
 def test_WalletManager(docker, request, devices_filled_data_folder, device_manager):
